requirements.txt
brainflow
hydra-core
pytorch-lightning
numpy
matplotlib
fastparquet
